+++ CHANGELOG.markdown
@@ -6,6 +6,8 @@
 * Removed `Indexing` and `IndexingResult` from `Control.Lens.Internal`.
 * Removed some of the more redundant combinators from `Control.Lens.Plated`.
 * Moved `partsOf`, `holesOf`, and `elementOf` into `Control.Lens.Traversal`.
+* Renamed `query` to `peruse` and `queries` to `peruses`. These are much less contentious names,
+  both contain `use` in their name for analogy to `use` and `uses` and the word is about reading.
 
 3.3
 ---